refactor(core): move DTCG value types to core with RealisedToken
